[Bridge] organized Monolog/HttpKernel bridge classes
