+++ src/Symfony/Bridge/Monolog/HttpKernel/Handler/DebugHandler.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bridge\Monolog\Handler;
+namespace Symfony\Bridge\Monolog\HttpKernel\Handler;
 
 use Monolog\Logger;
 use Monolog\Handler\TestHandler;
