+++ src/Symfony/Bridge/Monolog/HttpKernel/Handler/FirePHPHandler.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bridge\Monolog\Handler;
+namespace Symfony\Bridge\Monolog\HttpKernel\Handler;
 
 use Monolog\Handler\FirePHPHandler as BaseFirePHPHandler;
 use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
