+++ src/Symfony/Bundle/MonologBundle/Resources/config/monolog.xml
@@ -5,7 +5,7 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="monolog.logger.class">Symfony\Bridge\Monolog\Logger</parameter>
+        <parameter key="monolog.logger.class">Symfony\Bridge\Monolog\HttpKernel\Logger</parameter>
         <parameter key="monolog.handler.stream.class">Monolog\Handler\StreamHandler</parameter>
         <parameter key="monolog.handler.fingers_crossed.class">Monolog\Handler\FingersCrossedHandler</parameter>
         <parameter key="monolog.handler.group.class">Monolog\Handler\GroupHandler</parameter>
@@ -14,8 +14,8 @@
         <parameter key="monolog.handler.syslog.class">Monolog\Handler\SyslogHandler</parameter>
         <parameter key="monolog.handler.null.class">Monolog\Handler\NullHandler</parameter>
         <parameter key="monolog.handler.test.class">Monolog\Handler\TestHandler</parameter>
-        <parameter key="monolog.handler.firephp.class">Symfony\Bridge\Monolog\Handler\FirePHPHandler</parameter>
-        <parameter key="monolog.handler.debug.class">Symfony\Bridge\Monolog\Handler\DebugHandler</parameter>
+        <parameter key="monolog.handler.firephp.class">Symfony\Bridge\Monolog\HttpKernel\Handler\FirePHPHandler</parameter>
+        <parameter key="monolog.handler.debug.class">Symfony\Bridge\Monolog\HttpKernel\Handler\DebugHandler</parameter>
         <parameter key="monolog.handler.swift_mailer.class">Monolog\Handler\SwiftMailerHandler</parameter>
         <parameter key="monolog.handler.native_mailer.class">Monolog\Handler\NativeMailerHandler</parameter>
     </parameters>
